Use correct type for diskid. The *_DISKID defines are larger than a 32bit signed integer.

+++ main/songs.c
@@ -247,7 +247,7 @@ void songs_pause_resume(void)
 // 0 otherwise
 int songs_haved2_cd()
 {
-	int discid;
+	unsigned long discid;
 
 	if (GameCfg.OrigTrackOrder)
 		return 1;
